fix 64 to 32 bit clang conversion warning in src/core/lib

+++ src/core/lib/surface/channel.h
@@ -145,7 +145,7 @@ class Channel : public RefCounted<Channel>,
 
   int TestOnlyRegisteredCalls() {
     MutexLock lock(&registration_table_.mu);
-    return registration_table_.map.size();
+    return static_cast<int>(registration_table_.map.size());
   }
 
   int TestOnlyRegistrationAttempts() {
